+++ src/decoder/sync.rs
@@ -89,7 +89,7 @@ pub struct Start {
     pub address: u64,
 }
 
-impl<U> Decode<U> for Start {
+impl<U> Decode<'_, '_, U> for Start {
     fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
         let branch = decoder.read_bit()?;
         let ctx = Context::decode(decoder)?;
@@ -121,7 +121,7 @@ pub struct Trap {
     pub info: trap::Info,
 }
 
-impl<U> Decode<U> for Trap {
+impl<U> Decode<'_, '_, U> for Trap {
     fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
         let branch = decoder.read_bit()?;
         let ctx = Context::decode(decoder)?;
@@ -156,7 +156,7 @@ pub struct Context {
     pub context: Option<u64>,
 }
 
-impl<U> Decode<U> for Context {
+impl<U> Decode<'_, '_, U> for Context {
     fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
         let privilege = decoder
             .read_bits::<u8>(decoder.field_widths.privilege.get())?
@@ -194,7 +194,7 @@ pub struct Support<I = unit::ReferenceIOptions, D = unit::ReferenceDOptions> {
     pub doptions: D,
 }
 
-impl<U: Unit> Decode<U> for Support<U::IOptions, U::DOptions> {
+impl<U: Unit> Decode<'_, '_, U> for Support<U::IOptions, U::DOptions> {
     fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
         let ienable = decoder.read_bit()?;
         let encoder_mode = decoder
@@ -233,7 +233,7 @@ pub enum QualStatus {
     EndedNtr,
 }
 
-impl<U> Decode<U> for QualStatus {
+impl<U> Decode<'_, '_, U> for QualStatus {
     fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
         Ok(match decoder.read_bits::<u8>(2)? {
             0b00 => QualStatus::NoChange,